+++ src/router/users/reset.ts
@@ -1,24 +1,22 @@
-import { domain, limitCl, mg, mgDomain, usersCl, verificationCl } from "../../common";
+import { usersCl, verificationCl } from "../../common";
 import hash from "hash.js";
 import { Static, Type } from "@sinclair/typebox";
 import { ajv } from "../../lib/ajv";
 import User from "../../models/user";
-import Limit from "../../models/limit";
 import bcrypt from "bcrypt";
 import { createToken } from "../../lib/auth/createtoken";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
-import { randomBytes } from "crypto";
 
 export default (
     fastify: FastifyInstance,
-    opts: FastifyPluginOptions,
+    _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) => {
     const schema = Type.Object(
         {
             email: Type.String({ format: "email" }),
-            token: Type.Optional(Type.String({ minLength: 30, maxLength: 30 })),
-            pwd: Type.Optional(Type.RegEx(/^[a-f0-9]{64}$/i)),
+            code: Type.String({ minLength: 30, maxLength: 30 }),
+            pwd: Type.RegEx(/^[a-f0-9]{64}$/i),
         },
         { additionalProperties: false }
     );
@@ -29,87 +27,37 @@ export default (
             if (!ajv.validate(schema, req.body))
                 return res.code(400).send({ error: "Bad request." });
 
-            const { email, token, pwd } = req.body;
+            const { email, code, pwd } = req.body;
 
             const hashedEmail = hash.sha256().update(email).digest("hex");
 
-            const userData = (await usersCl.findOne({ email: hashedEmail })) as User;
-            if (!userData) return res.code(404).send({ error: "User not found." });
-
-            if (token && pwd) {
-                if (
-                    !(await verificationCl.findOne({
-                        type: "reset",
-                        email: hashedEmail,
-                        code: token,
-                    }))
-                )
-                    return res.code(401).send({
-                        error: "Token incorrect, or expired, or you have not requested reset password.",
-                    });
-
-                const userData = (await usersCl.findOne({ email: hashedEmail })) as User;
-
-                if (!userData) return res.code(404).send({ error: "User not found." });
-
-                await usersCl.updateOne(
-                    { email: hashedEmail },
-                    { $set: { pwd: bcrypt.hashSync(pwd, 10) } }
-                );
-
-                await verificationCl.deleteOne({
+            if (
+                !(await verificationCl.findOne({
                     type: "reset",
                     email: hashedEmail,
-                    code: token,
-                });
-
-                return res.send({
-                    token: createToken(
-                        userData.id,
-                        userData.name,
-                        userData.sex,
-                        userData.role
-                    ),
-                });
-            }
-
-            if (
-                (await limitCl.countDocuments({ type: "reset", email: hashedEmail })) >= 2
+                    code,
+                }))
             )
-                return res.code(429).send({
-                    error: "You can only request reset password 2 times a day.",
+                return res.code(401).send({
+                    error: "Token incorrect, or expired, or you have not requested reset password.",
                 });
 
-            const verificationCode = randomBytes(15).toString("hex");
+            const user = (await usersCl.findOne({ email: hashedEmail })) as User;
 
-            const reset = {
-                from: `Metahkg support <support@${mgDomain}>`,
-                to: req.body.email,
-                subject: "Metahkg - Reset Password",
-                text: `Reset your password with the following link:
-    https://${domain}/users/reset?code=${encodeURIComponent(
-                    verificationCode
-                )}&email=${encodeURIComponent(req.body.email)}
+            if (!user) return res.code(404).send({ error: "User not found." });
 
-    Alternatively, use this code at https://${domain}/reset :
-    ${verificationCode}`,
-            };
+            await usersCl.updateOne(
+                { email: hashedEmail },
+                { $set: { pwd: bcrypt.hashSync(pwd, 10) } }
+            );
 
-            await mg.messages().send(reset);
-
-            await verificationCl.insertOne({
+            await verificationCl.deleteOne({
                 type: "reset",
-                code: verificationCode,
-                email: userData.email,
+                email: hashedEmail,
+                code: code,
             });
 
-            await limitCl.insertOne({
-                type: "reset",
-                email: userData.email,
-                createdAt: new Date(),
-            } as Limit);
-
-            res.send({ response: "ok" });
+            return res.send({ token: createToken(user) });
         }
     );
     done();